+++ src/Library/math/Quantization.cpp
@@ -7,5 +7,5 @@ double quantization(double continuous_num, double resolution) {
 
 float quantization_f(double continuous_num, double resolution) {
   int bin_num = (int)((double)continuous_num / resolution);
-  return (float)bin_num * resolution;
+  return (float)bin_num * (float)resolution;
 }
